feat: refactored e2e tests
