+++ testdata/mock_relays.json
@@ -1,0 +1,92 @@
+[
+  {
+    "applicationPublicKey": "12345efc21889053171849c02dc71c4859b113c8b7b66dd2fbde268381e07a7c",
+    "nodePublicKey": "test_node_pub_key_02fbcfbad0777942c1da5425bf0105546e7e7f53fffad9",
+    "method": "eth_chainId",
+    "blockchain": "49",
+    "blockchainSubdomain": "fantom-mainnet",
+    "origin": "https://app.test1.io",
+    "elapsedTime": 0.16475
+  },
+  {
+    "applicationPublicKey": "12345e4c4c130d41268513268a376ed2204104b2ac4498a35d3fe26450460fb7",
+    "nodePublicKey": "test_node_pub_key_29f42f262efa514fbe93af1e963b40c5cddb2200dd9e0a",
+    "method": "eth_getBalance",
+    "blockchain": "49",
+    "blockchainSubdomain": "fantom-mainnet",
+    "origin": "https://app.test1.io",
+    "elapsedTime": 0.20045
+  },
+  {
+    "applicationPublicKey": "12345244e2ede722d756188316196fbf1018dbec087f6caee76bb4bc2861d46c",
+    "nodePublicKey": "test_node_pub_key_abfe693936a467d31e3246e32939ec04f5509315f39ef1",
+    "method": "eth_getBlockByNumber",
+    "blockchain": "49",
+    "blockchainSubdomain": "fantom-mainnet",
+    "origin": "https://app.test1.io",
+    "elapsedTime": 0.0813666666666667
+  },
+  {
+    "applicationPublicKey": "12345166969ae8263902693c0fc1d3569207a4f6f380d3d4d514e7fd819a69d4",
+    "nodePublicKey": "test_node_pub_key_277ed47d71a69e1aef94f8338f859af46d57840060f17d",
+    "method": "eth_getCode",
+    "blockchain": "3",
+    "blockchainSubdomain": "avax-mainnet",
+    "origin": "https://app.test2.io",
+    "elapsedTime": 0.15785
+  },
+  {
+    "applicationPublicKey": "12345fe0fc623aff8bba4ba1984e0c521c08874c9519cc6dcd518a15dd241f53",
+    "nodePublicKey": "test_node_pub_key_f64fe10e37173c0e7117490291c11e81d8851175e7ad4d",
+    "method": "eth_getTransactionCount",
+    "blockchain": "3",
+    "blockchainSubdomain": "avax-mainnet",
+    "origin": "https://app.test2.io",
+    "elapsedTime": 0.054675
+  },
+  {
+    "applicationPublicKey": "12345d9470aef46d0ebd3cb3076f3a5a3228c650e374c69b3665b0e5b0017a69",
+    "nodePublicKey": "test_node_pub_key_498b2b95ba2db09a263314e5fe23ce94fbd555d23578c1",
+    "method": "eth_getStorageAt",
+    "blockchain": "3",
+    "blockchainSubdomain": "avax-mainnet",
+    "origin": "https://app.test2.io",
+    "elapsedTime": 0.1093
+  },
+  {
+    "applicationPublicKey": "123455d57036c3bb5bd0de0319771cfdb3b2a28d4d64e33a3a23e6dcd6057f55",
+    "nodePublicKey": "test_node_pub_key_66b14f98b70d7e77572a6c72db611ebe28cf75d9cf542f",
+    "method": "eth_blockNumber",
+    "blockchain": "4",
+    "blockchainSubdomain": "bsc-mainnet",
+    "origin": "https://app.test3.io",
+    "elapsedTime": 0.2205
+  },
+  {
+    "applicationPublicKey": "12345f9fe17a1f6aaca7fe9df6499e569ae2f4910708f636beb1efa20cebda4d",
+    "nodePublicKey": "test_node_pub_key_b5c0fd8910bb7a121686865407e1a3ba40dcfb700d5e87",
+    "method": "eth_getBalance",
+    "blockchain": "4",
+    "blockchainSubdomain": "bsc-mainnet",
+    "origin": "https://app.test3.io",
+    "elapsedTime": 0.0932
+  },
+  {
+    "applicationPublicKey": "12345019c3f109073c77cd6d8bca9d1ff21b1ad5328ba04a5a610ba1bd72e1c5",
+    "nodePublicKey": "test_node_pub_key_124d8dcca024f05e01fb56c91800480926d62da44b0eee",
+    "method": "eth_blockNumber",
+    "blockchain": "4",
+    "blockchainSubdomain": "bsc-mainnet",
+    "origin": "https://app.test3.io",
+    "elapsedTime": 0.1162
+  },
+  {
+    "applicationPublicKey": "1234546a9d0c7d69ac65ffcd508b068ea2651e5d4bd5f4760bd2643584b9ef6d",
+    "nodePublicKey": "test_node_pub_key_dbd1105f79c1c40c2fc8d8749ea5760700cff6a0504016",
+    "method": "eth_blockNumber",
+    "blockchain": "49",
+    "blockchainSubdomain": "fantom-mainnet",
+    "origin": "https://app.test1.io",
+    "elapsedTime": 0.0814
+  }
+]
