+++ tests/e2e_test.go
@@ -1,3 +1,5 @@
+//go:build tests
+
 package tests
 
 import (
@@ -11,120 +13,123 @@ import (
 	"github.com/stretchr/testify/require"
 )
 
-func Test_RelayMeter_Collector_E2E(t *testing.T) {
-	if testing.Short() {
-		t.Skip("skipping end to end test")
-	}
-
-	c := require.New(t)
-
-	/* Create Test Influx Client */
-	client, orgID := initTestInfluxClient(c)
-	tasksAPI := client.TasksAPI()
-	writeAPI := client.WriteAPI(influxOptions.Org, mainBucket)
-
-	resetInfluxBuckets(client, orgID, c)
-	tasks := initInfluxTasks(tasksAPI, orgID, c)
-
-	/* Create Test Client to Verify InfluxDB contents */
-	testInfluxClient := db.NewInfluxDBSource(influxOptions)
-
-	tests := []struct {
-		name                  string
-		numberOfRelays        int
-		expectedDailyCounts   map[time.Time]map[string]api.RelayCounts
-		expectedHourlyLatency map[string]float64
-		err                   error
-	}{
-		{
-			name:           "Should collect a set number of relays from Influx",
-			numberOfRelays: 100_000,
-			expectedDailyCounts: map[time.Time]map[string]api.RelayCounts{
-				today: {
-					testRelays[0].applicationPublicKey: {Success: 8888, Failure: 1112},
-					testRelays[1].applicationPublicKey: {Success: 8889, Failure: 1111},
-					testRelays[2].applicationPublicKey: {Success: 8889, Failure: 1111},
-					testRelays[3].applicationPublicKey: {Success: 8889, Failure: 1111},
-					testRelays[4].applicationPublicKey: {Success: 8889, Failure: 1111},
-					testRelays[5].applicationPublicKey: {Success: 8889, Failure: 1111},
-					testRelays[6].applicationPublicKey: {Success: 8889, Failure: 1111},
-					testRelays[7].applicationPublicKey: {Success: 8889, Failure: 1111},
-					testRelays[8].applicationPublicKey: {Success: 8889, Failure: 1111},
-					testRelays[9].applicationPublicKey: {Success: 8888, Failure: 1111},
-				},
-			},
-			expectedHourlyLatency: map[string]float64{
-				testRelays[0].applicationPublicKey: 0.16475,
-				testRelays[1].applicationPublicKey: 0.20045,
-				testRelays[2].applicationPublicKey: 0.08137,
-				testRelays[3].applicationPublicKey: 0.15785,
-				testRelays[4].applicationPublicKey: 0.05467,
-				testRelays[5].applicationPublicKey: 0.1093,
-				testRelays[6].applicationPublicKey: 0.2205,
-				testRelays[7].applicationPublicKey: 0.0932,
-				testRelays[8].applicationPublicKey: 0.1162,
-				testRelays[9].applicationPublicKey: 0.0814,
-			},
-			err: nil,
-		},
-	}
-
-	for _, test := range tests {
-		t.Run(test.name, func(t *testing.T) {
-			/* Populate Relays in Influx DB */
-			populateInfluxRelays(writeAPI, today, test.numberOfRelays)
-			time.Sleep(1 * time.Second)
-			for _, task := range tasks {
-				_, err := tasksAPI.RunManually(ctx, task)
-				c.NoError(err)
-				time.Sleep(10 * time.Second) // Wait for task to complete
-			}
-
-			/* Verify Results from Influx Using Collector Influx Methods */
-			dailyCounts, err := testInfluxClient.DailyCounts(today, today.AddDate(0, 0, 1))
-			c.NoError(err)
-			totalSuccess, totalFailure := 0, 0
-			for _, count := range dailyCounts[today] {
-				totalSuccess += int(count.Success)
-				totalFailure += int(count.Failure)
-			}
-			// One relay missed due to collection interval between buckets - applies only to test
-			c.Equal(test.numberOfRelays-1, totalSuccess+totalFailure)
-			c.Equal(test.expectedDailyCounts, dailyCounts)
-
-			todaysCounts, err := testInfluxClient.TodaysCounts()
-			c.NoError(err)
-			for i, count := range todaysCounts {
-				c.NotEmpty(count.Success)
-				c.NotEmpty(count.Failure)
-				// Count will be for an incomplete day so less relays than Daily Count
-				c.LessOrEqual(count.Success, test.expectedDailyCounts[today][i].Success)
-				c.LessOrEqual(count.Failure, test.expectedDailyCounts[today][i].Failure)
-			}
-
-			todaysCountsPerOrigin, err := testInfluxClient.TodaysCountsPerOrigin()
-			c.NoError(err)
-			for origin, countPerOrigin := range todaysCountsPerOrigin {
-				// Daily Count by Origin query does not record failures
-				c.NotEmpty(countPerOrigin.Success)
-				c.Contains([]string{"https://app.test1.io", "https://app.test2.io", "https://app.test3.io"}, origin)
-			}
-
-			todaysLatency, err := testInfluxClient.TodaysLatency()
-			c.NoError(err)
-			for app, latencies := range todaysLatency {
-				for _, hourlyLatency := range latencies {
-					c.NotEmpty(hourlyLatency)
-					if hourlyLatency.Latency != 0 {
-						c.Equal(test.expectedHourlyLatency[app], hourlyLatency.Latency)
-					}
-				}
-			}
-		})
-	}
-
-	client.Close()
+var testClientOptions = TestClientOptions{
+	InfluxDBOptions: db.InfluxDBOptions{
+		URL:                 "http://localhost:8086",
+		Token:               "mytoken",
+		Org:                 "myorg",
+		DailyBucket:         "mainnetRelayApp1d",
+		CurrentBucket:       "mainnetRelayApp10m",
+		DailyOriginBucket:   "mainnetOrigin1d",
+		CurrentOriginBucket: "mainnetOrigin60m",
+	},
+	mainBucket:        "mainnetRelay",
+	main1mBucket:      "mainnetRelayApp1m",
+	phdBaseURL:        "http://localhost:8090",
+	phdAPIKey:         "test_api_key_6789",
+	testUserID:        "12345678fgte0db3b6c63124",
+	relayMeterBaseURL: "http://localhost:9898",
 }
+
+// func Test_RelayMeter_Collector_E2E(t *testing.T) {
+// 	if testing.Short() {
+// 		t.Skip("skipping end to end test")
+// 	}
+
+// 	c := require.New(t)
+
+// 	tc, err := NewTestClient(testClientOptions)
+// 	c.NoError(err)
+
+// 	tests := []struct {
+// 		name                  string
+// 		numberOfRelays        int
+// 		expectedDailyCounts   map[time.Time]map[string]api.RelayCounts
+// 		expectedHourlyLatency map[string]float64
+// 		err                   error
+// 	}{
+// 		{
+// 			name:           "Should collect a set number of relays from Influx",
+// 			numberOfRelays: 100_000,
+// 			expectedDailyCounts: map[time.Time]map[string]api.RelayCounts{
+// 				today: {
+// 					tc.TestRelays[0].ApplicationPublicKey: {Success: 8888, Failure: 1112},
+// 					tc.TestRelays[1].ApplicationPublicKey: {Success: 8889, Failure: 1111},
+// 					tc.TestRelays[2].ApplicationPublicKey: {Success: 8889, Failure: 1111},
+// 					tc.TestRelays[3].ApplicationPublicKey: {Success: 8889, Failure: 1111},
+// 					tc.TestRelays[4].ApplicationPublicKey: {Success: 8889, Failure: 1111},
+// 					tc.TestRelays[5].ApplicationPublicKey: {Success: 8889, Failure: 1111},
+// 					tc.TestRelays[6].ApplicationPublicKey: {Success: 8889, Failure: 1111},
+// 					tc.TestRelays[7].ApplicationPublicKey: {Success: 8889, Failure: 1111},
+// 					tc.TestRelays[8].ApplicationPublicKey: {Success: 8889, Failure: 1111},
+// 					tc.TestRelays[9].ApplicationPublicKey: {Success: 8888, Failure: 1111},
+// 				},
+// 			},
+// 			expectedHourlyLatency: map[string]float64{
+// 				tc.TestRelays[0].ApplicationPublicKey: 0.16475,
+// 				tc.TestRelays[1].ApplicationPublicKey: 0.20045,
+// 				tc.TestRelays[2].ApplicationPublicKey: 0.08137,
+// 				tc.TestRelays[3].ApplicationPublicKey: 0.15785,
+// 				tc.TestRelays[4].ApplicationPublicKey: 0.05467,
+// 				tc.TestRelays[5].ApplicationPublicKey: 0.1093,
+// 				tc.TestRelays[6].ApplicationPublicKey: 0.2205,
+// 				tc.TestRelays[7].ApplicationPublicKey: 0.0932,
+// 				tc.TestRelays[8].ApplicationPublicKey: 0.1162,
+// 				tc.TestRelays[9].ApplicationPublicKey: 0.0814,
+// 			},
+// 			err: nil,
+// 		},
+// 	}
+
+// 	for _, test := range tests {
+// 		t.Run(test.name, func(t *testing.T) {
+// 			/* Populate Relays in Influx DB */
+// 			tc.PopulateInfluxRelays(today, test.numberOfRelays)
+// 			tc.RunInfluxTasks()
+
+// 			/* Verify Results from Influx Using Collector Influx Methods */
+// 			dailyCounts, err := tc.Source.DailyCounts(today, today.AddDate(0, 0, 1))
+// 			c.NoError(err)
+// 			totalSuccess, totalFailure := 0, 0
+// 			for _, count := range dailyCounts[today] {
+// 				totalSuccess += int(count.Success)
+// 				totalFailure += int(count.Failure)
+// 			}
+// 			// One relay missed due to collection interval between buckets - applies only to test
+// 			c.Equal(test.numberOfRelays-1, totalSuccess+totalFailure)
+// 			c.Equal(test.expectedDailyCounts, dailyCounts)
+
+// 			todaysCounts, err := tc.Source.TodaysCounts()
+// 			c.NoError(err)
+// 			for i, count := range todaysCounts {
+// 				c.NotEmpty(count.Success)
+// 				c.NotEmpty(count.Failure)
+// 				// Count will be for an incomplete day so less relays than Daily Count
+// 				c.LessOrEqual(count.Success, test.expectedDailyCounts[today][i].Success)
+// 				c.LessOrEqual(count.Failure, test.expectedDailyCounts[today][i].Failure)
+// 			}
+
+// 			todaysCountsPerOrigin, err := tc.Source.TodaysCountsPerOrigin()
+// 			c.NoError(err)
+// 			for origin, countPerOrigin := range todaysCountsPerOrigin {
+// 				// Daily Count by Origin query does not record failures
+// 				c.NotEmpty(countPerOrigin.Success)
+// 				c.Contains([]string{"https://app.test1.io", "https://app.test2.io", "https://app.test3.io"}, origin)
+// 			}
+
+// 			todaysLatency, err := tc.Source.TodaysLatency()
+// 			c.NoError(err)
+// 			for app, latencies := range todaysLatency {
+// 				for _, hourlyLatency := range latencies {
+// 					c.NotEmpty(hourlyLatency)
+// 					if hourlyLatency.Latency != 0 {
+// 						c.Equal(test.expectedHourlyLatency[app], hourlyLatency.Latency)
+// 					}
+// 				}
+// 			}
+// 		})
+// 	}
+// }
 
 func Test_RelayMeter_APIServer_E2E(t *testing.T) {
 	if testing.Short() {
@@ -133,11 +138,13 @@ func Test_RelayMeter_APIServer_E2E(t *testing.T) {
 
 	c := require.New(t)
 
-	/* Initialize PHD Data */
-	err := populatePocketHTTPDB()
+	/* Initialize Test Client */
+	tc, err := NewTestClient(testClientOptions)
 	c.NoError(err)
 
-	time.Sleep(30 * time.Second) // Wait for collector to run and write to Postgres
+	/* Initialize PHD Data */
+	err = tc.PopulatePocketHTTPDB()
+	c.NoError(err)
 
 	tests := []struct {
 		name string
@@ -145,26 +152,28 @@ func Test_RelayMeter_APIServer_E2E(t *testing.T) {
 		appPubKey,
 		userID,
 		origin string
+		numberOfRelays        int
 		expectedHourlyLatency map[string]float64
 		err                   error
 	}{
 		{
-			name:      "Should return relays from the API server",
-			date:      today,
-			appPubKey: testRelays[0].applicationPublicKey,
-			userID:    testUserID,
-			origin:    testRelays[0].origin,
+			name:           "Should return relays from the API server",
+			date:           tc.startOfDay,
+			appPubKey:      tc.TestRelays[0].ApplicationPublicKey,
+			userID:         tc.options.testUserID,
+			origin:         tc.TestRelays[0].Origin,
+			numberOfRelays: 100_000,
 			expectedHourlyLatency: map[string]float64{
-				testRelays[0].applicationPublicKey: 0.16475,
-				testRelays[1].applicationPublicKey: 0.20045,
-				testRelays[2].applicationPublicKey: 0.08137,
-				testRelays[3].applicationPublicKey: 0.15785,
-				testRelays[4].applicationPublicKey: 0.05467,
-				testRelays[5].applicationPublicKey: 0.1093,
-				testRelays[6].applicationPublicKey: 0.2205,
-				testRelays[7].applicationPublicKey: 0.0932,
-				testRelays[8].applicationPublicKey: 0.1162,
-				testRelays[9].applicationPublicKey: 0.0814,
+				tc.TestRelays[0].ApplicationPublicKey: 0.16475,
+				tc.TestRelays[1].ApplicationPublicKey: 0.20045,
+				tc.TestRelays[2].ApplicationPublicKey: 0.08137,
+				tc.TestRelays[3].ApplicationPublicKey: 0.15785,
+				tc.TestRelays[4].ApplicationPublicKey: 0.05467,
+				tc.TestRelays[5].ApplicationPublicKey: 0.1093,
+				tc.TestRelays[6].ApplicationPublicKey: 0.2205,
+				tc.TestRelays[7].ApplicationPublicKey: 0.0932,
+				tc.TestRelays[8].ApplicationPublicKey: 0.1162,
+				tc.TestRelays[9].ApplicationPublicKey: 0.0814,
 			},
 			err: nil,
 		},
@@ -172,11 +181,19 @@ func Test_RelayMeter_APIServer_E2E(t *testing.T) {
 
 	for _, test := range tests {
 		t.Run(test.name, func(t *testing.T) {
+			/* Populate Relays in Influx DB */
+			tc.PopulateInfluxRelays(test.date, test.numberOfRelays)
+			time.Sleep(1 * time.Second)
+			/* Manually run the Influx tasks to populate time scale buckets from main bucket */
+			tc.RunInfluxTasks()
+
+			time.Sleep(30 * time.Second) // Wait for collector to run and write to Postgres
+
 			/* Test API Server */
 			dateParams := fmt.Sprintf("?from=%s&to=%s", test.date.Format(time.RFC3339), test.date.Format(time.RFC3339))
 
 			/* /relays */
-			allRelays, err := get[api.TotalRelaysResponse](relayMeterBaseURL, "v0/relays", "", dateParams)
+			allRelays, err := get[api.TotalRelaysResponse](tc.options.relayMeterBaseURL, "v0/relays", "", dateParams, "", tc.httpClient)
 			c.Equal(test.err, err)
 			c.NotEmpty(allRelays.Count.Success)
 			c.NotEmpty(allRelays.Count.Failure)
@@ -184,7 +201,7 @@ func Test_RelayMeter_APIServer_E2E(t *testing.T) {
 			c.Equal(allRelays.To, test.date.AddDate(0, 0, 1))
 
 			/* /relays/apps */
-			allAppsRelays, err := get[[]api.AppRelaysResponse](relayMeterBaseURL, "v0/relays/apps", "", dateParams)
+			allAppsRelays, err := get[[]api.AppRelaysResponse](tc.options.relayMeterBaseURL, "v0/relays/apps", "", dateParams, "", tc.httpClient)
 			c.Equal(test.err, err)
 			for _, appRelays := range allAppsRelays {
 				c.Len(appRelays.Application, 64)
@@ -195,7 +212,7 @@ func Test_RelayMeter_APIServer_E2E(t *testing.T) {
 			}
 
 			/* /relays/apps/{APP_PUB_KEY} */
-			appRelays, err := get[api.AppRelaysResponse](relayMeterBaseURL, "v0/relays/apps", test.appPubKey, dateParams)
+			appRelays, err := get[api.AppRelaysResponse](tc.options.relayMeterBaseURL, "v0/relays/apps", test.appPubKey, dateParams, "", tc.httpClient)
 			c.Equal(test.err, err)
 			c.Len(appRelays.Application, 64)
 			c.NotEmpty(appRelays.Count.Success)
@@ -204,7 +221,7 @@ func Test_RelayMeter_APIServer_E2E(t *testing.T) {
 			c.Equal(appRelays.To, test.date.AddDate(0, 0, 1))
 
 			/* /relays/users/{USER_ID} */
-			userRelays, err := get[api.UserRelaysResponse](relayMeterBaseURL, "v0/relays/users", test.userID, dateParams)
+			userRelays, err := get[api.UserRelaysResponse](tc.options.relayMeterBaseURL, "v0/relays/users", test.userID, dateParams, "", tc.httpClient)
 			c.Equal(test.err, err)
 			c.Len(userRelays.User, 24)
 			c.Len(userRelays.Applications, 10)
@@ -215,7 +232,7 @@ func Test_RelayMeter_APIServer_E2E(t *testing.T) {
 			c.Equal(userRelays.To, test.date.AddDate(0, 0, 1))
 
 			/* /relays/endpoints */
-			allEndpointsRelays, err := get[[]api.LoadBalancerRelaysResponse](relayMeterBaseURL, "v0/relays/endpoints", "", dateParams)
+			allEndpointsRelays, err := get[[]api.LoadBalancerRelaysResponse](tc.options.relayMeterBaseURL, "v0/relays/endpoints", "", dateParams, "", tc.httpClient)
 			c.Equal(test.err, err)
 			for _, endpointRelays := range allEndpointsRelays {
 				c.Len(endpointRelays.Endpoint, 24)
@@ -228,7 +245,7 @@ func Test_RelayMeter_APIServer_E2E(t *testing.T) {
 			}
 
 			/* /relays/endpoints/{ENDPOINT_ID} */
-			endpointRelays, err := get[api.LoadBalancerRelaysResponse](relayMeterBaseURL, "v0/relays/endpoints", allEndpointsRelays[0].Endpoint, dateParams)
+			endpointRelays, err := get[api.LoadBalancerRelaysResponse](tc.options.relayMeterBaseURL, "v0/relays/endpoints", allEndpointsRelays[0].Endpoint, dateParams, "", tc.httpClient)
 			c.Equal(test.err, err)
 			c.Len(endpointRelays.Endpoint, 24)
 			c.Len(endpointRelays.Applications, 1)
@@ -239,7 +256,7 @@ func Test_RelayMeter_APIServer_E2E(t *testing.T) {
 			c.Equal(endpointRelays.To, test.date.AddDate(0, 0, 1))
 
 			/* /relays/origin-classification */
-			allOriginRelays, err := get[[]api.OriginClassificationsResponse](relayMeterBaseURL, "v0/relays/origin-classification", "", dateParams)
+			allOriginRelays, err := get[[]api.OriginClassificationsResponse](tc.options.relayMeterBaseURL, "v0/relays/origin-classification", "", dateParams, "", tc.httpClient)
 			c.Equal(test.err, err)
 			for _, originRelays := range allOriginRelays {
 				c.Len(originRelays.Origin, 20)
@@ -251,7 +268,7 @@ func Test_RelayMeter_APIServer_E2E(t *testing.T) {
 			/* /relays/origin-classification/{ORIGIN} */
 			url, err := url.Parse(test.origin)
 			c.Equal(test.err, err)
-			originRelays, err := get[api.OriginClassificationsResponse](relayMeterBaseURL, "v0/relays/origin-classification", url.Host, dateParams)
+			originRelays, err := get[api.OriginClassificationsResponse](tc.options.relayMeterBaseURL, "v0/relays/origin-classification", url.Host, dateParams, "", tc.httpClient)
 			c.Equal(test.err, err)
 			c.Equal(url.Host, originRelays.Origin)
 			c.Len(originRelays.Origin, 12)
@@ -260,7 +277,7 @@ func Test_RelayMeter_APIServer_E2E(t *testing.T) {
 			c.Equal(originRelays.To, test.date.AddDate(0, 0, 1))
 
 			/* /latency/apps */
-			allAppLatencies, err := get[[]api.AppLatencyResponse](relayMeterBaseURL, "v0/latency/apps", "", dateParams)
+			allAppLatencies, err := get[[]api.AppLatencyResponse](tc.options.relayMeterBaseURL, "v0/latency/apps", "", dateParams, "", tc.httpClient)
 			c.Equal(test.err, err)
 			for _, appLatency := range allAppLatencies {
 				c.Len(appLatency.DailyLatency, 24)
@@ -276,7 +293,7 @@ func Test_RelayMeter_APIServer_E2E(t *testing.T) {
 			}
 
 			/* /latency/apps/{APP_PUB_KEY} */
-			appLatency, err := get[api.AppLatencyResponse](relayMeterBaseURL, "v0/latency/apps", test.appPubKey, dateParams)
+			appLatency, err := get[api.AppLatencyResponse](tc.options.relayMeterBaseURL, "v0/latency/apps", test.appPubKey, dateParams, "", tc.httpClient)
 			c.Equal(test.err, err)
 			c.Len(appLatency.DailyLatency, 24)
 			for _, hourlyLatency := range appLatency.DailyLatency {
